+++ CHANGELOG.md
@@ -15,6 +15,7 @@ any parts of the framework not mentioned in the documentation should generally b
 * Fixed invalid relationship pointer in error objects when field naming formatting is used.
 * Properly resolved related resource type when nested source field is defined.
 * Prevented overwriting of pointer in custom error object
+* Adhered to field naming format setting when generating OpenAPI schema.
 
 ### Added